Add files via upload
For this assignment, I made a simple iOS application that introduces a student. 

Features: input fields for first name, last name, college, major, switch for year in school, on switch for more pets wanted, increment/decrement for number of pets owned, and field for type of pets. Introduce student button presents this information in an alert style box. 

Bug: there is one issue with the text for type of pet -- I made a field for type of pet that is inputed in the message but if the type of pet field has no input the message just makes it empty so instead of "0 pets" it will just say "0" (from "I have 0 pets" to "I have 0") . In the next build I hope to implement a default message incase the pets text field is empty.